Put MumbleAPI in separate file and rename src/mumble/MumbleAPI to not have a name conflict
