+++ src/mumble/API.cpp
@@ -1,9 +1,10 @@
-// Copyright 2005-2019 The Mumble Developers. All rights reserved.
+// Copyright 2019-2020 The Mumble Developers. All rights reserved.
 // Use of this source code is governed by a BSD-style license
 // that can be found in the LICENSE file at the root of the
 // Mumble source tree or at <https://www.mumble.info/LICENSE>.
 
-#include "MumbleAPI.h"
+#include "API.h"
+#include "PluginComponents.h"
 #include "ClientUser.h"
 #include "Channel.h"
 #include "ServerHandler.h"
