+++ src/mumble/PluginManager.cpp
@@ -1,4 +1,4 @@
-// Copyright 2005-2019 The Mumble Developers. All rights reserved.
+// Copyright 2019-2020 The Mumble Developers. All rights reserved.
 // Use of this source code is governed by a BSD-style license
 // that can be found in the LICENSE file at the root of the
 // Mumble source tree or at <https://www.mumble.info/LICENSE>.
@@ -26,7 +26,7 @@
 #include "Log.h"
 #include "ProcessResolver.h"
 #include "ServerHandler.h"
-#include "MumbleAPI.h"
+#include "API.h"
 #include "PluginUpdater.h"
 
 #include <memory>
